melhorias na UX formularios e validacoes backend

+++ src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(request, response) {
@@ -13,6 +14,11 @@ class ContactController {
   async show(request, response) {
     //  Obter UM registro
     const { id } = request.params;
+
+    // validando se o UUID esta em formato correto
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Contact not found (invalid contact id)' });
+    }
 
     const contact = await ContactsRepository.findById(id);
 
@@ -33,16 +39,24 @@ class ContactController {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category id' });
+    }
 
-    // verificando se o email já está cadastrado
-    if (contactExists) {
-      return response.status(400).json({ error: 'This e-mail already in use' });
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+      // verificando se o email já está cadastrado
+      if (contactExists) {
+        return response.status(400).json({ error: 'This e-mail already in use' });
+      }
     }
 
     // caso não esteja cadastrado, vai para tela de cadastro
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
     response.status(201).json(contact);
@@ -55,23 +69,35 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
-    const contactExists = await ContactsRepository.findById(id);
-    if (!contactExists) {
-      return response.status(404).json({ error: 'Contact not found' });
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Contact not found (invalid contact id)' });
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category id' });
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
+    const contactExists = await ContactsRepository.findById(id);
+    if (!contactExists) {
+      return response.status(404).json({ error: 'Contact not found' });
+    }
 
-    // Verificar se o email existe, e verificar se o id encontrado é dirente do id que será alterado
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      return response.status(400).json({ error: 'This e-mail is already use' });
+    if (email) {
+      // Verificar se o email existe, e verificar se id encontrado é dirente do id que será alterado
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(400).json({ error: 'This e-mail is already use' });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
-      name, email, phone, category_id,
+      name,
+      email,
+      phone,
+      category_id: category_id || null, // caso o category_id nao for preenchido, retorna null
     });
 
     response.json(contact);
@@ -80,6 +106,10 @@ class ContactController {
   async delete(request, response) {
     //  Deletar um registro
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Contact not found (invalid contact id)' });
+    }
 
     await ContactsRepository.delete(id);
     // 204: No Content
